+++ sickrage/providers/torrent/filelist.py
@@ -62,8 +62,9 @@ class FileListProvider(TorrentProvider):
             "password": self.password
         }
 
-        response = sickrage.app.wsession.post(self.urls["login"], data=login_params).text
-        if not response:
+        try:
+            response = sickrage.app.wsession.post(self.urls["login"], data=login_params).text
+        except Exception:
             sickrage.app.log.warning("Unable to connect to provider")
             return False