+++ sickrage/providers/torrent/immortalseed.py
@@ -79,8 +79,9 @@ class ImmortalseedProvider(TorrentProvider):
             'password': self.password,
         }
 
-        response = sickrage.app.wsession.post(self.urls['login'], data=login_params).text
-        if not response:
+        try:
+            response = sickrage.app.wsession.post(self.urls['login'], data=login_params).text
+        except Exception:
             sickrage.app.log.warning("Unable to connect to provider")
             return False