+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 2e52131 - 2017-11-11: Added YGG torrent provider 
+- * 7336553 - 2017-11-11: refactored providers 
+- * aa6d45f - 2017-11-11: Added YGG torrent provider 
 - * 846f1d5 - 2017-11-11: Added YGG torrent provider 
 - * 7c3be4f - 2017-11-11: Added YGG torrent provider 
 - * f2a792d - 2017-11-11: Added Subscene subtitle provider